chore: MVC 구조를 위한 상수 분리 및 폴더 추가
