+++ src/constant/OutputMessage.js
@@ -1,0 +1,13 @@
+const MESSAGE = {
+    RACING_CHAMPION: '최종 우승자 : ',
+    RACING_CAR_INPUT: '경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)',
+    RACING_COUNT: '시도할 횟수는 몇 회인가요?',
+    RACING_PLAY: '실행 결과'
+};
+
+const ERROR_MESSAGE = {
+    INPUT_LENGTH_ERROR: '[ERROR] 입력한 자동차 이름이 5자 이상입니다.',
+    INPUT_NUMBER_ERROR: '[ERROR] 입력한 값이 숫자가 아닙니다.'
+}
+
+export { MESSAGE, ERROR_MESSAGE }
