+++ src/model/Comm.js
@@ -1,15 +1,5 @@
 import { Console } from "@woowacourse/mission-utils";
-
-const MESSAGE = {
-    RACING_CHAMPION: '최종 우승자 : ',
-    RACING_CAR_INPUT: '경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)',
-    RACING_COUNT: '시도할 횟수는 몇 회인가요?',
-    RACING_PLAY: '실행 결과'
-};
-
-const ERROR_MESSAGE = {
-    INPUT_ERROR: '[ERROR] 숫자가 잘못된 형식입니다.',
-}
+import { MESSAGE, ERROR_MESSAGE } from "../constant/OutputMessage";
 
 class Comm{
 
@@ -52,7 +42,7 @@ class Comm{
     racingCarInputValutate = (carName) => {
         return carName.reduce((arr, cur, idx) => {
             if(cur.length > 5){
-                throw Error(`${ERROR_MESSAGE.INPUT_ERROR}`);
+                throw Error(`${ERROR_MESSAGE.INPUT_LENGTH_ERROR}`);
             }
             return carName;
         }, "");
